+++ tests/unit_tests/unitTestAddOns/baseTestSetupMixins.py
@@ -551,6 +551,7 @@ class App(QObject):
     mwGlob = {'modelDir': 'tests/workDir/model',
               'imageDir': 'tests/workDir/image',
               'dataDir': 'tests/workDir/data',
+              'workDir': 'tests/workDir/data',
               'configDir': 'tests/workDir/config',
               }
     uiWindows = {}