+++ docs/source/changelog.rst
@@ -24,6 +24,8 @@ Beta version of MW4
 - add: auto abort imaging when camera device is disconnected
 - add: missing cursor in virtual keypad window
 - add: support for keyboard usage in virtual keypad window
+- add: screenshot as PNG save for actual window with key F5
+- add: screenshots as PNG save for all open windows with key F6
 - improved: flexible satellite handling when mount not connected
 - improved: show selected satellite name in satellite windows title
 - improved: 3D simulator drawing